pyproject.toml
[project]
name = "institutional-defi-platform-api"
version = "0.1.0"
description = "Unified institutional DeFi platform API"
requires-python = ">=3.11"

dependencies = [
    # FastAPI core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "pydantic>=2.9.1",
    "pydantic-settings>=2.5.2",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",

    # Database
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.2",

    # Data processing
    "pyyaml>=6.0.1",
    "pandas>=2.0.0",
    "networkx>=3.0",

    # Search
    "rank-bm25>=0.2.2",

    # HTTP client
    "httpx>=0.27.2",

    # Async tasks
    "celery>=5.4.0",
    "redis>=5.1.0",

    # Observability
    "structlog>=24.1.0",
    "python-json-logger>=2.0.7",

    # Security
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
blockchain = [
    "web3>=7.2.0",
]
llm = [
    "anthropic>=0.34.0",
]
ml = [
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
]
temporal = [
    "temporalio>=1.9.0",
]
telemetry = [
    "opentelemetry-api>=1.23.0",
    "opentelemetry-sdk>=1.23.0",
    "opentelemetry-instrumentation-fastapi>=0.44b0",
    "prometheus-client>=0.20.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]
all = [
    "institutional-defi-platform-api[blockchain,llm,ml,temporal,telemetry,dev]",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]
